+++ core/src/main/kotlin/shoyoream/server/shoyoreamapplication/core/domain/order/exception/OrderErrorType.kt
@@ -3,7 +3,9 @@ package shoyoream.server.shoyoreamapplication.core.domain.order.exception
 import shoyoream.server.shoyoreamapplication.core.common.exception.ErrorType
 
 enum class OrderErrorType : ErrorType {
-    NOT_FOUND_ORDER
+    NOT_FOUND_ORDER,
+    NOT_FOUND_STOCKS,
+    ALREADY_IN_ORDER
     ;
 
     override fun getErrorType() = this.name
