+++ core/src/main/kotlin/shoyoream/server/shoyoreamapplication/core/infra/config/RedisConfig.kt
@@ -1,5 +1,8 @@
 package shoyoream.server.shoyoreamapplication.core.infra.config
 
+import org.redisson.Redisson
+import org.redisson.api.RedissonClient
+import org.redisson.config.Config
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
@@ -18,18 +21,15 @@ class RedisConfig(
     @Value("\${spring.data.redis.host}") private val host: String,
     @Value("\${spring.data.redis.port}") private val port: Int
 ) {
+    companion object {
+        private const val REDISSON_HOST_PREFIX = "redis://"
+    }
+
     @Bean
     fun redisConnectionFactory(): LettuceConnectionFactory {
         val redisStandaloneConfiguration = RedisStandaloneConfiguration(host, port)
         return LettuceConnectionFactory(redisStandaloneConfiguration)
     }
-
-//    @Bean
-//    fun redisTemplate(): RedisTemplate<*, *> {
-//        val redisTemplate = RedisTemplate<ByteArray, ByteArray>()
-//        redisTemplate.connectionFactory = redisConnectionFactory()
-//        return redisTemplate
-//    }
 
     @Bean
     fun redisTemplate(): RedisTemplate<Any, Any> {
@@ -38,5 +38,14 @@ class RedisConfig(
         template.valueSerializer = Jackson2JsonRedisSerializer(Any::class.java)
         template.connectionFactory = redisConnectionFactory()
         return template
+    }
+
+    @Bean
+    fun redissonClient(): RedissonClient {
+        val config = Config()
+        config.useSingleServer()
+            .setAddress("$REDISSON_HOST_PREFIX$host:$port")
+
+        return Redisson.create(config)
     }
 }
