--- core/src/main/kotlin/shoyoream/server/shoyoreamapplication/core/infra/config/RedissonConfig.kt
@@ -1,27 +1,0 @@
-package shoyoream.server.shoyoreamapplication.core.infra.config
-
-import org.redisson.Redisson
-import org.redisson.api.RedissonClient
-import org.redisson.config.Config
-import org.springframework.beans.factory.annotation.Value
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-
-@Configuration
-class RedissonConfig(
-    @Value("\${spring.data.redis.host}") private val host: String,
-    @Value("\${spring.data.redis.port}") private val port: Int
-) {
-    companion object {
-        private const val REDISSON_HOST_PREFIX = "redis://"
-    }
-
-    @Bean
-    fun redissonClient(): RedissonClient {
-        val config = Config()
-        config.useSingleServer()
-            .setAddress("$REDISSON_HOST_PREFIX$host:$port")
-
-        return Redisson.create(config)
-    }
-}
