+++ core/src/main/resources/application-redis.yml
@@ -1,5 +1,5 @@
 spring:
     data:
         redis:
-            host: localhost
+            host: redis
             port: 6379
