+++ order-application/src/main/resources/graphql/schema.graphqls
@@ -24,16 +24,12 @@ enum GoodsSize {
 
 input OrderInput {
     goodsId: String!
-    goodsType: GoodsType!
-    goodsSize: GoodsSize!
     buyerId: Int!
     stocksId: String!
 }
 
 input StockInput {
     goodsId: String!
-    goodsType: GoodsType!
-    goodsSize: GoodsSize!
     sellerId: Int!
     price: Float!
 }
